requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
croniter==1.4.1
dateutils==0.6.12
deepdiff==6.3.1
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.103.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
h5py==3.9.0
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lightning==2.0.7
lightning-cloud==0.5.37
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.25.2
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.24.2
psutil==5.9.5
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.7
pytz==2023.3
PyYAML==6.0.1
readchar==4.0.5
requests==2.31.0
rich==13.5.2
scikit-learn==1.3.0
scipy==1.11.2
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
starsessions==1.3.0
threadpoolctl==3.2.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchinfo==1.8.0
torchmetrics==1.1.1
torchvision==0.12.0+cu113
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
wandb==0.15.9
wcwidth==0.2.6
websocket-client==1.6.2
websockets==11.0.3
yarl==1.9.2
